fix(qsrv): store +putorder at full int64 width with i64::MIN sentinel bump (BRIDGE-RS-2026-05-28-45)

+++ crates/epics-bridge-rs/src/qsrv/group.rs
@@ -970,13 +970,13 @@ impl GroupChannel {
         // position (first), matching the absent-putOrder ordering. Before
         // this fix the `filter_map` dropped them, so a proc-only save/apply
         // hook without `+putorder` silently never ran.
-        let mut ordered: Vec<(&GroupMember, i32)> = self
+        let mut ordered: Vec<(&GroupMember, i64)> = self
             .def
             .members
             .iter()
             .filter_map(|m| match m.put_order {
                 Some(po) => Some((m, po)),
-                None if m.mapping == FieldMapping::Proc => Some((m, i32::MIN)),
+                None if m.mapping == FieldMapping::Proc => Some((m, i64::MIN)),
                 None => None,
             })
             .collect();
